churn_project/requirements.txt
Django
djangorestframework
django-filter
mlflow
numpy
pandas
psycopg2_binary
scikit_learn
celery
django-redis
sqlalchemy
matplotlib
seaborn
joblib
python-dotenv
tqdm
plotly
ipykernel
jupyter
django-cors-headers==4.3.1
requests==2.31.0